WIP: Not working
The generated code (tsc) outputs ES6 style modules. Node 10.14 does not support this without the experimental flag.
Jest tests won't run/pass because of it.

Need to figure out a solution for it. Might have to bring in Webpack as part of the process.

After further experimentation, I finally found: https://stackoverflow.com/a/40357573/4410648

This provided the solution required. I don't yet fully understand the "export = FOO;" syntax, but... ¯\_(ツ)_/¯